POS_App/requirements.txt
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pytest>=7.0.0
